Create Autonomous Driving Simulation
autonomous driving simulation in Java can be an ambitious and interesting project. However, fully implementing an autonomous driving system from scratch in Java requires a considerable amount of resources and knowledge, especially for tasks like object detection, planning, and control. For simplicity, we'll build a simulation that mimics some of the basic aspects of autonomous driving, such as lane-following using a simple Proportional-Derivative (PD) controller and simple obstacle detection.